requirements.txt
psycopg2-binary
fastapi[all]
pydantic
uvicorn
python-dotenv
sqlalchemy